[Debug] Add replica_placement

+++ dqn.py
@@ -200,7 +200,7 @@ class DQN():
       for k,v in zip(["w1","b1","w2","b2"], [W1,b1,W2,b2]):
          print("Variable: ", k)
          print("Shape: ", v.shape)
-      self.state_input = tf.compat.v1.placeholder("float",[None,self.state_dim+add])
+      self.state_input = tf.compat.v1.placeholder("float64",[None,self.state_dim+add])
       print(" self.state_input Shape: ", self.state_input.dtype)
       h_layer = tf.nn.relu(tf.matmul(self.state_input,W1) + b1)
       self.Q_value = tf.matmul(h_layer,W2) + b2
